ml/requirements.txt
python-socketio
uvicorn
numpy
pandas
scikit-learn
python-dotenv